MSE800-PSE/week1/rentflex-full-release/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
aiomysql
python-dotenv
alembic
passlib[bcrypt]
python-jose
cryptography
mysql-connector-python

